+++ shared/utils.js
@@ -7,7 +7,7 @@ module.exports = {
   setDefaults() {
     this.options.stage =
       _.get(this, 'options.stage') || _.get(this, 'serverless.service.provider.stage') || 'dev';
-    this.options.runtime = _.get(this, 'options.runtime') || 'nodejs8';
+    this.options.runtime = _.get(this, 'options.runtime') || 'nodejs10';
 
     // serverless framework is hard-coding us-east-1 region from aws
     // this is temporary fix for multiple regions